chore: Moving EF interface and WithParam to embeddings
