+++ src/main/java/tech/amikos/chromadb/Client.java
@@ -1,6 +1,7 @@
 package tech.amikos.chromadb;
 
 import com.google.gson.internal.LinkedTreeMap;
+import tech.amikos.chromadb.embeddings.EmbeddingFunction;
 import tech.amikos.chromadb.handler.ApiClient;
 import tech.amikos.chromadb.handler.ApiException;
 import tech.amikos.chromadb.handler.DefaultApi;
