+++ src/main/java/tech/amikos/chromadb/embeddings/EmbeddingFunction.java
@@ -1,7 +1,9 @@
-package tech.amikos.chromadb;
+package tech.amikos.chromadb.embeddings;
+
+import tech.amikos.chromadb.EFException;
+import tech.amikos.chromadb.Embedding;
 
 import java.util.List;
-import java.util.Map;
 
 public interface EmbeddingFunction {
