+++ src/main/java/tech/amikos/chromadb/embeddings/ollama/OllamaEmbeddingFunction.java
@@ -3,6 +3,8 @@ package tech.amikos.chromadb.embeddings.ollama;
 import com.google.gson.Gson;
 import okhttp3.*;
 import tech.amikos.chromadb.*;
+import tech.amikos.chromadb.embeddings.EmbeddingFunction;
+import tech.amikos.chromadb.embeddings.WithParam;
 
 import java.io.IOException;
 import java.util.*;
