+++ src/main/java/tech/amikos/chromadb/embeddings/openai/OpenAIEmbeddingFunction.java
@@ -6,6 +6,8 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
 import tech.amikos.chromadb.*;
+import tech.amikos.chromadb.embeddings.EmbeddingFunction;
+import tech.amikos.chromadb.embeddings.WithParam;
 
 import java.io.IOException;
 import java.util.Arrays;
